fix: minor style-refactor according comments
