+++ src/components/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Header from './Header/Header';
-import FilterBar from './FilterBar/FilterBar';
-import Footer from './Footer/Footer';
+import Header from './Header';
+import FilterBar from './FilterBar';
+import Footer from './Footer';
 
 const App = () => (
   <>
