+++ src/components/FilterBar/_filterBar.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .filterBar {
   margin-inline: 50px;
