+++ src/components/FilterDropdown/_filterDropdown.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .filterDropdown {
   display: flex;
   align-items: center;
@@ -18,7 +20,7 @@
     padding-right: 20px;
     font-size: 16px;
     line-height: 19px;
-    color: white;
+    color: $white;
     background: transparent;
     text-transform: uppercase;
   }
