+++ src/components/Footer/_footer.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .footer {
   display: flex;
   justify-content: center;
