+++ src/components/Logo/_logo.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .logo {
   color: $red;
   font-size: 20px;
