+++ src/components/Logo/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import './_logo.scss';
+
 const Logo = () => (
   <span className="logo">
     <span>netflix</span>
