+++ src/components/MovieCard/index.tsx
