+++ src/components/SearchBar/_searchBar.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .searchBar {
   margin-left: 70px;
   backdrop-filter: blur(0);
