+++ src/components/TopBar/_topBar.scss
@@ -1,3 +1,5 @@
+@import '../../scss/variables';
+
 .topBar {
   display: flex;
   justify-content: space-between;
@@ -9,13 +11,16 @@
     padding: 5px 10px;
   }
 
-  &__link,
-  &__link:hover {
+  &__link {
     text-decoration: none;
     font-weight: 600;
     font-size: 20px;
     color: $red;
     line-height: 24px;
     text-transform: uppercase;
+
+    &:hover {
+      color: $red;
+    }
   }
 }
